requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
websocket-client==1.6.4
Werkzeug==3.0.0
loguru==0.7.2